requirements.txt
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
numpy==1.24.2
pandas==2.0.0
requests==2.28.2
